requirements.txt
openai>=1.0.0
pytest>=7.0.0
pymarkdownlnt>=0.9.34
